Reuse article page’s ReplyItem in reply page
