+++ redux/replyDetail.js
@@ -1,5 +1,6 @@
 import { createDuck } from 'redux-duck';
 import { fromJS } from 'immutable';
+import { waitForAuth } from './auth';
 import gql from '../util/gql';
 
 const { defineType, createAction, createReducer } = createDuck('replyDetail');
@@ -8,6 +9,7 @@ const { defineType, createAction, createReducer } = createDuck('replyDetail');
 //
 
 const LOAD = defineType('LOAD');
+const LOAD_AUTH = defineType('LOAD_AUTH');
 const SET_STATE = defineType('SET_STATE');
 const RESET = defineType('RESET');
 
@@ -25,15 +27,20 @@ export const load = id => dispatch => {
         versions(limit: 1) {
           type
           text
+          reference
           createdAt
         }
         replyConnections {
+          id
           article {
             id
             text
           }
           user {
             name
+          }
+          feedbacks {
+            score
           }
           status
           createdAt
@@ -46,16 +53,38 @@ export const load = id => dispatch => {
   });
 };
 
+export const loadAuth = id => dispatch => {
+  dispatch(setState({ key: 'isAuthLoading', value: true }));
+  return waitForAuth
+    .then(() =>
+      gql`
+        query($id: String!) {
+          GetReply(id: $id) {
+            replyConnections {
+              id
+              canUpdateStatus
+            }
+          }
+        }
+      `({ id })
+    )
+    .then(resp => {
+      dispatch(createAction(LOAD_AUTH)(resp.getIn(['data', 'GetReply'])));
+      dispatch(setState({ key: 'isAuthLoading', value: false }));
+    });
+};
+
 export const reset = () => createAction(RESET);
 
 // Reducer
 //
 
 const initialState = fromJS({
-  state: { isLoading: false },
+  state: { isLoading: false, isAuthLoading: false },
   data: {
     // data from server
     reply: null,
+    originalReplyConnection: null,
   },
 });
 
@@ -65,18 +94,34 @@ export default createReducer(
       state.setIn(['state', key], value),
 
     [LOAD]: (state, { payload }) => {
+      const originalReplyConnection = payload
+        .getIn(['replyConnections'])
+        .sortBy(item => item.get('createdAt'))
+        .first();
 
       return state
         .setIn(['data', 'reply'], payload)
-        .set('currentVersion', payload.getIn(['versions', 0]))
-        .set(
-          'originalArticle',
-          payload
-            .getIn(['replyConnections'])
-            .sortBy(item => item.get('createdAt'))
-            .first()
-            .get('article')
+        .setIn(
+          ['data', 'originalReplyConnection'],
+          originalReplyConnection.set('reply', payload)
         );
+    },
+
+    [LOAD_AUTH]: (state, { payload }) => {
+      // Write LOAD_AUTH results to existing replyConnections
+      //
+      const idAuthMap = payload.get('replyConnections').reduce((agg, conn) => {
+        agg[conn.get('id')] = conn;
+        return agg;
+      }, {});
+
+      const updateConnection = conn => conn.merge(idAuthMap[conn.get('id')]);
+
+      return state
+        .updateIn(['data', 'reply', 'replyConnections'], replyConnections =>
+          replyConnections.map(updateConnection)
+        )
+        .updateIn(['data', 'originalReplyConnection'], updateConnection);
     },
 
     [RESET]: state => state.set('data', initialState.get('data')),
